Refactor - create a display module which handles displaying the leaderboard scores and message, create an apiCalls module which handles all calls to the API
